+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## $v
 
 - Adding support for `_no_cache` attribute.
-- Bumping minimum required version of the WP Sharks Core.
 - Bumping minimum required WordPress version to v4.6.
+- Bumping minimum required version of the WP Sharks Core.
 - Adding support for `wp-markdown-extra` filter configuration.
 - Removing unnecessary nested shortcode callback handler. Starting with WP 4.6 the new `WP_Hook` class addresses the underlying problem this was solving. See [Issue #1](https://github.com/websharks/if-shortcode-pro/issues/1).
 - Enhancing security be removing `basename(__FILE__)` from direct access notices.
 - Removing unnecessary lite build variation from `.build.props` file.
+- Cleaned up the options page just a bit and corrected some typos.
 
 ## v160919.18816