+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v170407.77796-RC
 
 - Adding support for `_no_cache` attribute.
 - Adding support for `ua_is_engine="true|false"` attribute.